requirements.txt
discord.py>=2.5.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
